feat: ✨ Deploy from public-cloud-examples
