+++ solution/ai/deploy/YOLOV8/.dockerignore
@@ -1,0 +1,3 @@
+.dockerignore
+Dockerfile
+./runs/
