--- src/app/solution/.dockerignore
@@ -1,3 +1,0 @@
-.dockerignore
-Dockerfile
-./runs/
