--- src/app/solution/requirements.txt
@@ -1,3 +1,0 @@
-ultralytics==8.0.194 
-opencv-python-headless==4.8.1.78 
-streamlit==1.27.2
